# 8300184: Optimize ResourceHashtableBase::iterate_all using _number_of_entries