@dholmes-ora 2023-01-17
Nit: the bug synopsis says `iterate_all` but you actually changed `iterate`.