@navyxliu 2023-01-17
Hi, David, 

Thank you for reviewing this patch. 

> Nit: the bug synopsis says `iterate_all` but you actually changed `iterate`.

This is intentional. iterate_all() calls iterate() with a lambda '()->true'.  For iterate(), there are two cases. If the lambda returns constant value true, it mandates iterate() to traverse all elements. By contrast, user-defined lambdas could return false and stop iterating immediately. This patch only optimizes the former case. That's why I claim to optimize iterate_all rather than iterate.  

Sometimes, we need to traverse all elements. eg. we want to print out the contents, copy hashable or do statistics.

thanks,
--lx