@dholmes-ora 2023-01-18
> That's why I claim to optimize iterate_all rather than iterate.

A call to `iterate` can result in iterating all elements if the lambda does not find an element to trigger iteration to cease; in which case this optimisation also applies.